refactor(chunking): move DocumentChunker to vector_stores/
Chunking is invoked exclusively by ingestable vector stores during
ingest — sources don't see chunks, bindings don't see chunks. Move
the module next to its consumers so the dependency arrow points
inward (concrete vector store → shared vector_store utility) instead
of outward into dataset_types/.

The chunker docstring is also updated: it parses an IngestFile into
text chunks plus per-page images and hands off to whichever
IngestableVectorStore invoked it. Storage and retrieval stay with
the vector store.
