+++ backend/syft_space/components/vector_stores/chromadb_local/chromadb_vector_store.py
@@ -16,10 +16,6 @@ from typing import Any
 from loguru import logger
 from pydantic import ValidationError
 
-from syft_space.components.dataset_types.chunking import (
-    DocumentChunker,
-    build_image_urls,
-)
 from syft_space.components.shared.domain_types import (
     HealthcheckResponse,
     HealthcheckStatus,
@@ -37,6 +33,10 @@ from syft_space.components.vector_stores.chromadb_local.provisioner import (
 )
 from syft_space.components.vector_stores.chromadb_local.schemas import (
     ChromaDBLocalVectorStoreConfiguration,
+)
+from syft_space.components.vector_stores.chunking import (
+    DocumentChunker,
+    build_image_urls,
 )
 
 try:
