+++ backend/syft_space/components/vector_stores/chunking.py
@@ -1,8 +1,8 @@
 """Shared document chunking pipeline using Docling.
 
-Provides a unified parse → chunk → save images pipeline used by all
-dataset types (ChromaDB, Weaviate, etc.). Each dataset type handles
-its own DB-specific storage and retrieval.
+Parses an ``IngestFile`` into a sequence of text chunks plus
+per-page images, then leaves storage and retrieval to whichever
+``IngestableVectorStore`` invoked it.
 """
 
 import json
